extracted player and popup classes
